apps/iss/apps/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dateutil==2.8.2
email-validator==2.1.0
slowapi==0.1.9
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pillow==10.1.0
stripe==7.8.0
fastapi-mail==1.4.1

# Performance & Observability (P1)
fastapi-cache2==0.2.1
prometheus-fastapi-instrumentator==6.1.0
sentry-sdk[fastapi]==1.39.1

# Bando monitoring system
beautifulsoup4==4.12.2
apscheduler==3.10.4
lxml==4.9.3

# Testing dependencies
pytest-asyncio==0.21.1
aiosqlite==0.19.0
